// lib/acoes/comunicar_certificado_icpbrasil/Comunicar_Certificado_ICPBRASIL.d.ts
import { ModelComunicarCertificadoICPBRASIL } from "../../model/acoes/comunicar_certificado_icpbrasil/Comunicar_Certificado_ICPBRASIL";
export declare const comunicarCertificadoICPBRASIL: (_numeroSessao: number, _codigoAtivacao: string, _certificado: string) => Promise<ModelComunicarCertificadoICPBRASIL>;