// lib/acoes/configurar_interface_rede/Configurar_Interface_Rede.d.ts
import { ModelConfigurarInterfaceRedeDadosConfiguracao } from "../../model/acoes/configurar_interface_rede/Configurar_Interface_Rede_Dados_Configuracao";
import { ModelConfigurarInterfaceRede } from "../../model/acoes/configurar_interface_rede/Configurar_Interface_Rede";
export declare const configurarInterfaceRede: (_numeroSessao: number, _codigoAtivacao: string, _dadosConfiguracao: ModelConfigurarInterfaceRedeDadosConfiguracao) => Promise<ModelConfigurarInterfaceRede>;